+++ show-article.php
@@ -1,16 +1,20 @@
 <?php
+require_once __DIR__ . '/database/database.php';
+require_once __DIR__ . '/database/security.php';
 
-        /**
-         * @var ArticleDAO
-         */
+$currentUser = isLoggedIn();
 
-        // $pdo = require_once './database.php';
-        // $statement = $pdo->prepare('SELECT * FROM article where id=:id');
+/**
+ * @var ArticleDAO
+ */
 
-        // $filename = __DIR__ . "/data/articles.json";
-        // $articles = [];
+// $pdo = require_once './database.php';
+// $statement = $pdo->prepare('SELECT * FROM article where id=:id');
 
-        $articleDAO = require_once './database/models/ArticleDAO.php';
+// $filename = __DIR__ . "/data/articles.json";
+// $articles = [];
+
+$articleDAO = require_once './database/models/ArticleDAO.php';
 
 $_GET = filter_input_array(INPUT_GET, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
 $id = $_GET['id'] ?? '';
@@ -44,10 +48,13 @@ if (!$id) {
                 <h1 class="article-title"><?= $article['title'] ?></h1>
                 <div class="separator"></div>
                 <p class="article-content"><?= $article['content'] ?></p>
-                <div class="action">
-                    <a class="btn" href="/delete-article.php?id=<?= $article['id'] ?>">Supprimer</a>
-                    <a class="btn btn-primary" href="/form-article.php?id=<?= $article['id'] ?>">Éditer l'article</a>
-                </div>
+                <p class="article-author"><?= $article['firstname'] . ' ' . $article['lastname'] ?></p>
+                <?php if ($currentUser && $currentUser['id'] === $article['author']) : ?>
+                    <div class="action">
+                        <a class="btn" href="/delete-article.php?id=<?= $article['id'] ?>">Supprimer</a>
+                        <a class="btn btn-primary" href="/form-article.php?id=<?= $article['id'] ?>">Éditer l'article</a>
+                    </div>
+                <?php endif; ?>
             </div>
         </div>
         <?php require_once 'includes/footer.php' ?>
